Extract isAdjacent util for tests
